Fully Optimized Sorting Algorithms
