--- Heap Sort.c
@@ -1,50 +1,0 @@
-//Divide and Conquer Algorithm
-//Heap Sort
-
-#include <stdio.h>
-
-void heapify(int arr[], int n, int i) {
-    int largest = i, left = 2 * i + 1, right = 2 * i + 2;
-
-    if (left < n && arr[left] > arr[largest])
-        largest = left;
-    if (right < n && arr[right] > arr[largest])
-        largest = right;
-    
-    if (largest != i) {
-        int temp = arr[i];
-        arr[i] = arr[largest];
-        arr[largest] = temp;
-        heapify(arr, n, largest);
-    }
-}
-
-void heapSort(int arr[], int n) {
-    for (int i = n / 2 - 1; i >= 0; i--)
-        heapify(arr, n, i);
-    for (int i = n - 1; i > 0; i--) {
-        int temp = arr[0];
-        arr[0] = arr[i];
-        arr[i] = temp;
-        heapify(arr, i, 0);
-    }
-}
-
-int main() {
-    int arr[] = {4, 10, 3, 5, 1};
-    printf("The original array \n");
-    for (int i = 0; i < 5;i++){
-    	printf("%d ",arr[i]);
-    	
-	}
-	printf("\n");
-    int n = sizeof(arr) / sizeof(arr[0]);
-
-    heapSort(arr, n);
-
-    printf("Sorted array: ");
-    for (int i = 0; i < n; i++)
-        printf("%d ", arr[i]);
-        printf("\nProgram Executed By Aadarsha Jha(023BSCIT050)");
-    return 0;
-}
