--- Insertion Sort.c
@@ -1,27 +1,0 @@
-#include <stdio.h>
-
-void insertionSort(int arr[], int n) {
-    for (int i = 1; i < n; i++) {
-        int key = arr[i];
-        int j = i - 1;
-
-        while (j >= 0 && arr[j] > key) {
-            arr[j + 1] = arr[j];
-            j--;
-        }
-        arr[j + 1] = key;
-    }
-}
-
-int main() {
-    int arr[] = {12, 11, 13, 5, 6};
-    int n = sizeof(arr) / sizeof(arr[0]);
-
-    insertionSort(arr, n);
-
-    printf("Sorted array: ");
-    for (int i = 0; i < n; i++)
-        printf("%d ", arr[i]);
-        printf("\nProgram Executed By Aadarsha Jha(023BSCIT050)");
-    return 0;
-}
