+++ MergeSort.c
@@ -1,0 +1,64 @@
+//Divide and Conquer Algortihm
+//Merge Sort
+
+#include <stdio.h>
+#define N 10000
+
+void merge(int arr[], int left, int mid, int right) {
+    int n1 = mid - left + 1;
+    int n2 = right - mid;
+
+    int L[n1], R[n2];
+
+    for (int i = 0; i < n1; i++)
+        L[i] = arr[left + i];
+    for (int j = 0; j < n2; j++)
+        R[j] = arr[mid + 1 + j];
+
+    int i = 0, j = 0, k = left;
+    while (i < n1 && j < n2) {
+        if (L[i] <= R[j])
+            arr[k++] = L[i++];
+        else
+            arr[k++] = R[j++];
+    }
+
+    while (i < n1)
+        arr[k++] = L[i++];
+    while (j < n2)
+        arr[k++] = R[j++];
+}
+
+void mergeSort(int arr[], int left, int right) {
+    if (left < right) {
+        int mid = left + (right - left) / 2;
+        mergeSort(arr, left, mid);
+        mergeSort(arr, mid + 1, right);
+        merge(arr, left, mid, right);
+    }
+}
+
+int main() {
+    int arr[N];
+    int n,i;
+    printf("Enter the total no. of elements\n");
+    scanf_s("%d",&n);
+    printf("Enter %d elements\n",n);
+    for(i = 0  ;i < n; i++){
+        scanf_s("%d",&arr[i]);
+    }
+
+    printf("The original array \n");
+    for (int i = 0; i < n;i++){
+    	printf("%d ",arr[i]);
+    	
+	}
+    printf("\n");
+    mergeSort(arr, 0, n - 1);
+
+    printf("Sorted array: ");
+    for (int i = 0; i < n; i++)
+        printf("%d ", arr[i]);
+        printf("\nProgram Executed By Aadarsha Jha(023BSCIT050)");
+    return 0;
+}
