--- Selection Sort.c
@@ -1,28 +1,0 @@
-#include <stdio.h>
-
-void selectionSort(int arr[], int n) {
-    for (int i = 0; i < n - 1; i++) {
-        int minIndex = i;
-        for (int j = i + 1; j < n; j++) {
-            if (arr[j] < arr[minIndex])
-                minIndex = j;
-        }
-        // Swap
-        int temp = arr[minIndex];
-        arr[minIndex] = arr[i];
-        arr[i] = temp;
-    }
-}
-
-int main() {
-    int arr[] = {29, 10, 14, 37, 13};
-    int n = sizeof(arr) / sizeof(arr[0]);
-
-    selectionSort(arr, n);
-
-    printf("Sorted array: ");
-    for (int i = 0; i < n; i++)
-        printf("%d", arr[i]);
-        printf("\nProgram Executed By Aadarsha Jha(023BSCIT050)");
-    return 0;
-}
