--- Shell Sort.c
@@ -1,25 +1,0 @@
-#include <stdio.h>
-
-void shellSort(int arr[], int n) {
-    for (int gap = n / 2; gap > 0; gap /= 2) {
-        for (int i = gap; i < n; i++) {
-            int temp = arr[i], j;
-            for (j = i; j >= gap && arr[j - gap] > temp; j -= gap)
-                arr[j] = arr[j - gap];
-            arr[j] = temp;
-        }
-    }
-}
-
-int main() {
-    int arr[] = {12, 34, 54, 2, 3};
-    int n = sizeof(arr) / sizeof(arr[0]);
-
-    shellSort(arr, n);
-
-    printf("Sorted array: ");
-    for (int i = 0; i < n; i++)
-        printf("%d ", arr[i]);
-        printf("\nProgram Executed By Aadarsha Jha(023BSCIT050)");
-    return 0;
-}
